refactor: 특정 컴포넌트에서만 사용하는 타이핑을 `types.ts`가 아닌 각 컴포넌트에서 관리되도록 변경
