+++ src/components/Button/Button.types.ts
@@ -1,0 +1,2 @@
+export type ButtonVariant = 'primary' | 'text' | 'link';
+export type ButtonShape = 'default' | 'circle';
