+++ src/components/Card/Card.types.ts
@@ -1,0 +1,1 @@
+export type CardVariant = 'primary' | 'simple';
